// packages/sdk/src/views/models/timelineTypes.ts
export enum RiverTimelineEvent {
    ChannelCreate = 'm.channel.create',
    ChannelMessage = 'm.channel.message',
    ChannelMessageEncrypted = 'm.channel.encrypted',
    ChannelMessageEncryptedWithRef = 'm.channel.encrypted_with_ref',
    ChannelMessageMissing = 'm.channel.missing',
    ChannelProperties = 'm.channel.properties',
    EncryptedChannelProperties = 'm.channel.encrypted_properties',
    Fulfillment = 'm.fulfillment',
    Inception = 'm.inception', // TODO: would be great to name this after space / channel name
    InteractionRequest = 'm.interaction_request',
    InteractionRequestEncrypted = 'm.interaction_request.encrypted',
    InteractionResponse = 'm.interaction_response',
    KeySolicitation = 'm.key_solicitation',
    MemberBlockchainTransaction = 'm.member_blockchain_transaction',
    MiniblockHeader = 'm.miniblockheader',
    Pin = 'm.pin',
    Reaction = 'm.reaction',
    RedactedEvent = 'm.redacted_event',
    RedactionActionEvent = 'm.redaction_action',
    SpaceUpdateAutojoin = 'm.space.update_autojoin',
    SpaceUpdateHideUserJoinLeaves = 'm.space.update_channel_hide_user_join_leaves',
    SpaceImage = 'm.space.image',
    SpaceReview = 'm.space.review',
    StreamEncryptionAlgorithm = 'm.stream_encryption_algorithm',
    StreamMembership = 'm.stream_membership',
    TipEvent = 'm.tip_event',
    TokenTransfer = 'm.token_transfer',
    Unpin = 'm.unpin',
    UserBlockchainTransaction = 'm.user_blockchain_transaction',
    UserReceivedBlockchainTransaction = 'm.user_received_blockchain_transaction',
    ConversationSeed = 'm.conversation_seed',
    ConversationSeedResponse = 'm.conversation_seed_response',
    SetAppAddress = 'm.set_app_address',
    UnsetAppAddress = 'm.unset_app_address',
}
// ...
export interface TipEvent {
    kind: RiverTimelineEvent.TipEvent
    transaction: PlainMessage<BlockchainTransaction>
    tip: PlainMessage<BlockchainTransaction_Tip>
    transactionHash: string
    fromUserId: string
    refEventId: string
    toUserId: string
}
// ...
export interface RedactedEvent {
    kind: RiverTimelineEvent.RedactedEvent
    isAdminRedaction: boolean
}

// the event that redacted the original event
export interface RedactionActionEvent {
    kind: RiverTimelineEvent.RedactionActionEvent
    refEventId: string
    adminRedaction: boolean
}